refactor: 중복으로 정의한 타입 공통 타입으로 관리
