+++ frontend/src/pages/channel/types/Peer.d.ts
@@ -1,0 +1,12 @@
+export interface PeerConnection {
+  peerId: string;
+  peerNickname: string;
+  stream: MediaStream;
+  isHost: boolean;
+  reaction: string;
+}
+
+export interface PeerMediaStatus {
+  audio: boolean;
+  video: boolean;
+}
